Vendor openstacksdk delete_unattached_floating_ips
We're relying on this method to cleanup leaked floating ips for us.
Unfortunately, we suspect that it may have deleted floating ips that it
shouldn't have. In order to add more debugging around this process
vendor the code into zuul's openstackendpoint then update it to use our
internal connection client and add debug logging for each deleted
floating ip.

Hopefully this confirms that the floating ips are being deleted by this
routine and we get timestamps we can use to further track down the
underlying issue.

Change-Id: I519fa433ec7dd56645ca40f86731dae5216e2adf