cmd: catch and log all exceptions when exiting
Make sure the exit_handler call terminates even if an exception is
thrown while it is running.

Here is an example of the issue that happened during a graceful stop
of the executor (triggered by a rollout event on kubernetes) while the
zookeeper service was unavailable; the executor couldn't update its state on zk:

```
2025-11-21 08:54:51,983 DEBUG zuul.zk.base.ZooKeeperClient: ZooKeeper connection (session: 0x0): SUSPENDED
2025-11-21 08:55:23,917 INFO zuul.ExecutorServer: Stopping graceful
2025-11-21 08:55:23,917 DEBUG zuul.ExecutorServer: Pausing
2025-11-21 08:58:53,117 DEBUG zuul.zk.base.ZooKeeperClient: ZooKeeper connection (session: 0x0): LOST
Traceback (most recent call last):
  File "/usr/local/bin/zuul-executor", line 10, in <module>
    sys.exit(main())
    (..8<..)
    self.executor.graceful()
    (..8<..)
    kazoo.exceptions.SessionExpiredError
2025-11-21 08:58:53,159 DEBUG zuul.zk.base.ZooKeeperClient: ZooKeeper connection (session: 0x200000e99430001): CONNECTED
2025-11-21 08:58:53,160 INFO zuul.Component: Registering component in ZooKeeper /zuul/components/executor/zuul-executor-0.zuul-executor-headless.sf.svc.cluster.local
2025-11-21 08:58:53,207 INFO zuul.ComponentRegistry: Noticed new web component zuul-web-b99f5ff8-z2hsg0000000002
```

The uncaught exception cancels the graceful stop process, and the
executor pod never terminates as expected.

Since any exception encountered during the stopping process might leave
the zuul ensemble in a degraded or incorrect state, log the issue but
terminate anyway, as we assume stopping the component prevails.

Additionnally, use KazooRetry in the Zookeeper BaseComponent's __setattr__ method
to mitigate connection issues with Zookeeper that may be caused by a zk
server not being available (which is okay as long as a quorum is still established).

Change-Id: Ie4b5f4b16d19f913c5f53f81a675f4dde50a7189