Openstack: validate image checksum on upload
If the user supplied an md5sum or sha256sum with the image, then
instruct openstacksdk to validate that checksum after uploading.

According to the docs, the cloud computes this checksum, and if
they don't match, it will raise an exception (which we should
handle in the usual way and retry the upload).

Change-Id: I27106e23c406890a9a6a116894f9699ea47a020e