WIP Start sketching out implementation of web RBAC spec
There are a number of pieces to sort out when implementing the RBAC spec
from configuration loading to accurate application of rules on web
requests. Start an early sketch of what this might look like so that we
can identify any concerns with the approach before too much time is
spent committing to it.

Note this is completely non functional in this state, but I think it
gives us something concrete enough to assess the approach.

Depends-On: https://review.opendev.org/c/zuul/zuul/+/967938
Change-Id: Ibaaabcba5334a26f6e14a030b02f0f926207e15f