WIP Start sketching out implementation of web RBAC spec
There are a number of pieces to sort out when implementing the RBAC spec
from configuration loading to accurate application of rules on web
requests. Start an early sketch of what this might look like so that we
can identify any concerns with the approach before too much time is
spent committing to it.

To make this work we update the cherrpy check_root_auth and
check_tenant_auth decorators to run before the handler functions rather
than on_start_resource. The reaason for this is we need the JSON
parameters to be available to auth checking so that we can apply role
conditions to authenticated users to authorize them for the requested
action. We then combine the JSON request data with the url parameters to
process whether an authenticated user meets the rbac role conditions for
the actions.

This results in some compromises. The biggest of which is that if an
item isn't a request parameter then role conditions can't be applied to
it for that request. Requests like delete autohold only operate on
tenant and request_id so we can't filter by project or branch etc.

The upside to this approach is that authentication and authorization
checks are centralized and don't need bespoke code for each request to
handle request specific conditions. This greatly simplifies the
implementation which is a good thing when it comes to something like
authentication and authorization.

More discussion of this approach and its compromises was carried out in
I2b9bef2cc4bbc9c8cfed242d907064f17e6467ec which has since been squashed
into this change.

Note while there is testing that works this is currently an incomplete
implementation of the RBAC spec.

Co-Authored-By: James E. Blair <jim@acmegating.com>
Depends-On: https://review.opendev.org/c/zuul/zuul/+/967938
Change-Id: Ibaaabcba5334a26f6e14a030b02f0f926207e15f