Fix missing streaming logs from loop include_tasks
With Ansible 11 support in Zuul we no longer get the streaming output of
tasks that are executed via `include_tasks` in a loop.

The reason for that is the fix of an Ansible bug reported in
https://github.com/ansible/ansible/issues/82655 with the fix correctly
setting the `ansible_loop_var` and `ansible_index_var` variables
(only fixed in Ansible 10; not backported).

Since we have the module available we can also check the `task.loop`
variable analog to how this is already done in the callback plugin.

To account for empty loops we make those conditions more explicit by
checking for `not None`.

Change-Id: If86a6597f10383f2f6164ad3fed005684505b38b