Fix missing streaming logs from loop include_tasks
With Ansible 11 support in Zuul we no longer get the streaming output of
tasks that are executed via `include_tasks` in a loop.

The reason for that is the fix of an Ansible bug reported in
https://github.com/ansible/ansible/issues/82655 with the fix correctly
setting the `ansible_loop_var` and `ansible_index_var` variables
(only fixed in Ansible 10; not backported).

Historically it looks like we've used the presence of `ansible_loop_var`
as a task variable to detect loops and not setup streaming in those
cases.

Since this code path has been refactored to only perform the stream
setup for the *_shell and *_command modules, we no longer need to check
for loop tasks.

Change-Id: If86a6597f10383f2f6164ad3fed005684505b38b