Fix missing streaming logs from loop include_tasks
With Ansible 11 support in Zuul we no longer get the streaming output of
tasks that are executed via `include_tasks` in a loop.

The reason for that is the fix of an Ansible bug reported in
https://github.com/ansible/ansible/issues/82655 with the fix correctly
setting the `ansible_loop_var` and `ansible_index_var` variables
(only fixed in Ansible 10; not backported).

Since we only perform the stream setup for the *_shell and *_command
modules, it's unclear why this should be necessary.

From Ic1358d5b9f939549ffdd5d770fe4eafc047be6f1 it looks like there was a
problem where we would only get the output from the first item.

However, with the currently supported Ansible version this issue wasn't
reproducible anymore. This could also be due to more recent changes to
the streaming code that might have fixed this particular edge-case.

Change-Id: If86a6597f10383f2f6164ad3fed005684505b38b