Add reporter jobs
This adds reporter jobs as described in the spec.  The gerrit,
github, and gitlab drivers are all able to supply the necessary
merge sha information after a change merges in order to provide
that as input to reporter jobs.  Pagure does not (and I was
unable to find a way to get that information through the API),
so it is not implemnted for that driver and the limitation is
noted in the docs.

Change-Id: If81e2edf672ebc4187303986e9f27d22677cb38e