Fix promotion of changes behind and add test coverage
If change B is in a dependent pipeline and gets promoted its children
should only be promoted with it if they were previously in the pipeline
and meet pipeline requirements. Previously, this was not the case as
we ignored requirements when processing changes behind. Fix that.

Then to make the expected behaviors between pipeline requirements and
promotion more clear we add a number of test cases to cover a variety of
situations:

 1. Promotion when all changes meet the requirement
 2. Promotion when the promoted change has a child that does not meet
    the requirement
 3. Promotion when the promoted change has a parent that does not meet
    the requirement
 4. Promotion when the promoted change does not meet the requirement but
    its parent does.
 5. Promotion when the head of the queue is not related to the promoted
    changes and the promoted change does not meet requirements.

Co-Authored-By: James E. Blair <jim@acmegating.com>
Change-Id: I49c9791c2dacc6c101040db37b030123c3e37f76