Add tests for promotion in dependent pipelines with requirements
Pipeline requirements are evaluated when enqueuing changes due to
trigger events. Zuul users (admins) can also promote changes which is
similar to a dequeue then enqueue of the changes. Due to this
it is reasonable to expect pipeline requirements are evaluated when
promoting changes.

Add four new test cases that cover different scenarios for promotion of
a change in a dependent gate pipeline with a gerrit approval vote
requirement:
 1. Promotion when all changes meet the requirement
 2. Promotion when the promoted change has a child that does not meet
    the requirement
 3. Promotion when the promoted change has a parent that does not meet
    the requirement
 4. Promotion when the promoted change does not meet the requirement

Change-Id: I49c9791c2dacc6c101040db37b030123c3e37f76