Add NIZ provider attribute documentation
This adds documentation for all of the provider attributes for
all of the NIZ drivers.

Most of the text is based on the Nodepool docs where relevant,
with some changes.

The narrative documentation as well as the documentation for
the rest of the config objects and connection configuration
will be added in future changes.  Since this is still in
development, all the documentation pages are orphans (they
are not linked to in the index).

Change-Id: Ic4e04762d977a1ac1db4b55b3132b0b2a9eaf029
Depends-On: https://review.opendev.org/971480