AWS: make snapshot/image import/copy more resumable
If we have started an image or snapshot import task, and the
launcher crashes or is stopped, we may be able to simply find
that task in AWS and resume waiting for its completion.  This
can avoid errors if we try to import the same image again, and
can help speed recovery in case of a crash or stopped launcher.

Similarly, if we intend to copy an image, check to see if the
destination image already exists.

Change-Id: Ic60dacca76eaac96c01e75abe3b7ac4fb2d177cb